bug fix: long vs in conflict

+++ src/main/java/com/deallock/backend/repositories/DealWaitlistRepository.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DealWaitlistRepository extends JpaRepository<DealWaitlistEntry, Long> {
-    boolean existsByDealIdAndUserId(Long dealId, int userId);
+    boolean existsByDealIdAndUserId(Long dealId, long userId);
     long countByDealId(Long dealId);
-    Optional<DealWaitlistEntry> findFirstByDealIdAndUserId(Long dealId, int userId);
+    Optional<DealWaitlistEntry> findFirstByDealIdAndUserId(Long dealId, long userId);
 }
